+++ release-notes.md
@@ -1,7 +1,6 @@
 **1.2.4**
 
-* New API call /transactions/info
-
+* New API calls /transactions/info and /peers/connect
 * MVStore is used for state persistence
 * Fixed JSON style in API calls
 * PeersHttpService was rebaned to PeersApiRoute